Disallow dead code/unused imports by default in `wasmtime` crate
This PR is a change to the crate attributes of the `wasmtime` crate.
Quite some time ago when Wasmtime had far fewer crate features I added a
directive to allow dead code and unused imports when the `default`
feature was disabled. The original intention was to not worry too much
about unused imports and dead code in every possible configuration of
Wasmtime to ease the burden of dealing with unused imports or
compile-time conditionally used code.

The primary consequence of this change though is that `cargo build`
doesn't actually warn about dead code or unused imports in this
repository during development. The reason for this is that the `default`
feature is disabled by default and never actually enabled by any
dependency inside the repository. Instead each crate, like
`wasmtime-cli`, has its own `default` feature.

If I were to start from the problem of "I want dead code warnings by
default in this repository" there's two possible fixes:

1. Change `wasmtime-cli`'s `default` feature to include `wasmtime/default`.
2. Remove this `#[allow]` directive.

While (1) would indeed work this is an attempt to get (2) working to see
how bad the changes are necessary within the crate.

prtest:full