Cut down on serde requirements in `wasmtime` crate
Remove `derive(Deserialize, Serialize)` for some configuration-based
enums to ideally give us more flexibility in the future as to what these
are exactly called. I'm not sure of any internal users of these except
for one use due to #9811 which can be switched to using an option-based
parser to ensure the same syntax is accepted via `--config` and on the
CLI with a flag.