Start work towards always-warnings in `wasmtime`
This commit is an attempt to make progress after #10108. The goal of
this commit is to create a list of features to burn down over time and
to have `dead_code` and `unused_imports` lints on-by-default in
Wasmtime. The `feature = "default"` gate on this `allow` directive was
replaced with individual features. The hope is that as individual
features are removed from this list we can fix the resulting warnings
and the commit such a ratchet.