Start work towards always-warnings in wasmtime (#10131)
* Start work towards always-warnings in `wasmtime`

This commit is an attempt to make progress after #10108. The goal of
this commit is to create a list of features to burn down over time and
to have `dead_code` and `unused_imports` lints on-by-default in
Wasmtime. The `feature = "default"` gate on this `allow` directive was
replaced with individual features. The hope is that as individual
features are removed from this list we can fix the resulting warnings
and the commit such a ratchet.

* Burn down some features from allowing dead code

Remove the features from the crate and test compiling with all other
features to ensure that no warnings are issued.

* Run full tests in CI

prtest:full

* Fix warning condition

* Fix #[cfg] to fix warnings on iOS

Switch `target_os = "macos"` to `target_vendor = "apple"` to align
requirements.