Add a style guideline for conditional compilation
I've been doing a fair amount of work recently that's touched on
`#[cfg]` in many ways throughout Wasmtime. I've personally got opinions
about how to structure everything as well, and these opinions are not
always obvious to others or discerned from just reading snippets. To
assist with this I figured it would be nice to have a style guideline
for Wasmtime's conditionally compiled code explaining at least at a high
level what's going on and some rough basic principles.

I've attempted to give this a stab and have added a page to the
contributing documentation about the style guidelines for conditional
compilation. I'm sure I've forgotten something here but my hope is that
we can evolve this over time.