Fix OOM in fuzzing from using malloc memory (#10164)
Our `StoreLimits` implementation doesn't take into account the
reservation of linear memory that it can grow into. This works fine for
`mmap` since memory isn't committed, but it doesn't work in fuzzing for
malloc-based memories because the fuzzing harness just thinks a huge
allocation is being made and declares OOM. This is fixed in this commit
by ensuring that the `memory-reservation-for-growth` parameter is tuned
smaller-than-the-default-2G when malloc-based memories are used.