wasi-nn: pull test artifacts from new location (#10188)
Over in the [openvino-rs] repository, we removed large test artifacts
that required Git LFS and moved these artifacts to a new download site,
[download.01.org]. This broke Wasmtime's CI and was fixed in #9380 by
pinning to a commit version. This change adopts the [download.01.org]
site in Wasmtime's CI to fully cut the dependency to GitHub's LFS
bandwidth limits.

[openvino-rs]: https://github.com/intel/openvino-rs
[download.01.org]: https://download.01.org/openvinotoolkit/fixtures

prtest:full