Slightly simplify logic in `ref.test` translation
When the ref is null or i31, just use an `iconst` result value directly, rather
than reusing the result of the check for null or i31. Reusing the result doesn't
actually yield better code (equivalently good or worse depending on if null/i31
are allowed or not) and is somewhat subtle -- I have to stop and re-think
through its correctness each time I see it again -- so this change should be a
welcome improvement.

This does not change the logic of the emitted code, but does slightly change the
emitted code itself.