winch(aarch64): Correct treatment for stores and other trapping ops
This commit is one more in the series of executing spec tests for
aarch64.

It's mostly a small follow-up to https://github.com/bytecodealliance/wasmtime/pull/10146,
in which I omitted contextualizing the memory flags for stores as well
as ensuring that the SP is aligned when emitting other trapping
instructions like `checked_uadd`.