Update cargo-nextest used on CI (#10211)
This commit updates the version of `cargo nextest` used on CI from
0.9.67 to 0.9.88. It turns out that the nightly used in testing,
`nightly-2025-01-09`, cannot actually compile 0.9.67 due to a bug in the
`ahash` crate dependency. This never showed up prior on CI because we
cache the build of `cargo nextest` so the update of rustc to a new
nightly did not force a rebuild to happen. Now that the cache has been
removed the error is now showing up, so this will hopefully unblock the
queue.