aarch64: Add support for the `extr` instruction (#10229)
* aarch64: Add support for the `extr` instruction

This is pattern-matched from `bor` patterns of a specific shape. I found
this when doing some benchmarking of Wasmtime on aarch64 and I saw LLVM
generating this pattern but Wasmtime didn't. I didn't perform any
benchmarking between wasmtime/native though, so I'm just relying on this
reducing the number of instructions to probably be a wee bit faster.

* Review comments

* Fixing tests