asm: comprehend 1-, 2-, and 3-byte opcodes (#10232)
* asm: comprehend 1-, 2-, and 3-byte opcodes

How IA-32e describes and encodes opcodes is a bit confusing, but section
2.1.2 in the reference manual does a decent job explaining how the
various prefixes, escape bytes, and primary/secondary opcodes may be
used. This change uses this explanation to create what is essentially a
parser for `[u8]`; this allows us to compactly express instruction
definitions, e.g., `rex([0x66, 0x25])`. Improper use of this DSL syntax
should result in an immediate compilation error when attempting to
generate the assembler.

* Fix merge conflict