x64: Add support for the `shld` instruction (#10233)
This commit is similar to #10229 except that it uses the x64 `shld`
instruction. This instruction was not previously supported by Cranelift
so I opted to add support via the new `cranelift-assembler-x64` crate
instead of trying to fit it into the existing `emit.rs`/`MInst`
infrastructure. This was a good learning experience for myself and was
also fun to do!