asm: add implementations for more ALU operations (#10237)
* asm: add implementations for more ALU operations

This adds DSL definitions in the assembler _and_ uses the new
instructions for ISLE lowering for the following instructions:
- `add`
- `adc`
- `or`
- `sub`
- `sbb`
- `xor`

The original `AluRmiROpcode` variants are not yet gone; perhaps in a
future pass.

* review: add extra assert

* fix: bless new Cranelift test

* fix: bless Wasmtime disassembly

* review: add rules for smaller encodings

`cranelift-codegen` would check an immediate size and, if it fit in an
8-bit slot, would use the sign-extending version to encode an
instruction with a 1 byte immediate instead of a 4 byte immediate. This
change adds those extra lowering rules for the new assembler.