Introduce verification of integer address type widths. (#10209)
* cranelift/codegen/verifier: Add verification of pointer width (#10118)

This adds a check to load that confirms the pointer width is as expected
according to the target.

* cranelift/verifier: Add load pointer width verification.

Also clarify width of integer address type and unit tests that check
the new verifier rule.

* cranelift/filetests: Split out 64 and 32 bit loads.

Makes unit test pass with the verifier checking the load address
size.

* cranelift/verifier: Add address verification to more instructions.

Also adds unit tests to ensure problematic cases are detected.

* cranelift/verifier: Change wording, restructure arms.