x64: Refactor assembler ISLE constructors
This commit is spawned out of discussion between me and Andrew in
conjunction with the thoughts in #10238. The goal here is to pave a way
forward for various kinds of instructions in the future as well as give
access to more instructions today we already have formats for.

The major changes in this commit are:

* `Assembler*` types are gone from ISLE, except for immediates. Instead
  types like `Gpr` and `GprMem` are used instead.
* Rust-defined constructors for each instruction return `MInst` instead
  of implicitly performing an `emit` operation.
* Instructions with a read/write `GprMem` operand now generate two ISLE
  constructors instead of one. One constructor takes `Gpr` and returns
  `Gpr`, the other takes `Amode` and returns `SideEffectNoResult`.
* Generated ISLE constructors now match the SSA-like form of VCode/ISLE
  we already have. For example `AssemblerReadWriteGpr` is never used as
  a result, it's just `Gpr` instead. Conversions happen in Rust during
  construction of assembler instructions.

Using this new support various `x64_*_mem` instructions have now moved
over to the new assembler and using that instead. Looking to the future
this is intended to make it easier to generate constructors that return
`ProducesFlags` or `ConsumesFlags` such as `x64_adc` and friends. This
will require more refactoring to enable this but the goal is to move
roughly in such a direction.

I've attempted to make this abstract enough that it'll be relatively
easily extensible in the future to more ISLE constructors with minimal
changes, so some abstractions here may not be fully used just yet but
the hope is that they will be in the near future.