x64: ensure 8-bit sign extensions have higher priorities
In cases where an 8-bit immediate is used, x64 can emit a shorter
encoding and rely on the CPU to sign extend it. Several Cranelift tests
rely on this (as they should--this improves code size). The previous
commit breaks the previous priorities; this change restores them,
ensuring we use the smallest encoding when possible.