x64: rework new assembler rule priorities, remove old emission rules (#10260)
* x64: Use priority overlap in new instruction lowerings

Use overlapping priorities when the only difference is in the type being
matched on, and otherwise use priorities to disambiguate the arms of
`GprMemImm` and the various possibilities of extraction.

* x64: ensure 8-bit sign extensions have higher priorities

In cases where an 8-bit immediate is used, x64 can emit a shorter
encoding and rely on the CPU to sign extend it. Several Cranelift tests
rely on this (as they should--this improves code size). The previous
commit breaks the previous priorities; this change restores them,
ensuring we use the smallest encoding when possible.

* x64: remove old emission replaced by new assembler

We expect the new assembler rules to prevent us from reaching the old
emission rules. This change removes the old rules completely.

* x64: re-prioritize based review, add line breaks

* review: remove `is_mem` extractor for now

---------

Co-authored-by: Alex Crichton <alex@alexcrichton.com>