winch(aarch64): Correct treatment for stores and other trapping ops (#10201)
* winch(aarch64): Correct treatment for stores and other trapping ops

This commit is one more in the series of executing spec tests for
aarch64.

It's mostly a small follow-up to https://github.com/bytecodealliance/wasmtime/pull/10146,
in which I omitted contextualizing the memory flags for stores as well
as ensuring that the SP is aligned when emitting other trapping
instructions like `checked_uadd`.

* Add note around why SP alignment is needed in `checked_uadd`